Replace Postgres dependency with YAML flat files
The Postgres dependency complicates development and deployment of the
application. Given that while running most of the data is kept in memory,
that there are a relatively small number of cards, and that the relational
features of the database aren't in use, we can instead rely on flat files
and native Python data structures.